+++ miden-vm/tests/integration/operations/u32_ops/conversion_ops.rs
@@ -1,7 +1,8 @@
+use miden_core::field::PrimeField64;
 use miden_processor::ExecutionError;
 use miden_utils_testing::{
-    Felt, StarkField, U32_BOUND, WORD_SIZE, ZERO, build_op_test, expect_exec_error_matches,
-    proptest::prelude::*, rand::rand_value,
+    Felt, U32_BOUND, WORD_SIZE, build_op_test, expect_exec_error_matches, proptest::prelude::*,
+    rand::rand_value,
 };
 
 use super::{prop_randw, test_inputs_out_of_bounds};
@@ -103,10 +104,9 @@ fn u32assert_fail() {
 
     expect_exec_error_matches!(
         test,
-        ExecutionError::NotU32Values{ values, err_code, label: _, source_file: _ } if
+        ExecutionError::NotU32Values{ values, label: _, source_file: _ } if
             values.len() == 1 &&
-            values[0] == Felt::new(equal) &&
-            err_code == ZERO
+            values[0] == Felt::new(equal)
     );
 
     // --- test when a > 2^32 ---------------------------------------------------------------------
@@ -114,10 +114,9 @@ fn u32assert_fail() {
 
     expect_exec_error_matches!(
         test,
-        ExecutionError::NotU32Values{ values, err_code, label: _, source_file: _ } if
+        ExecutionError::NotU32Values{ values, label: _, source_file: _ } if
             values.len() == 1 &&
-            values[0] == Felt::new(larger) &&
-            err_code == ZERO
+            values[0] == Felt::new(larger)
     );
 }
 
@@ -148,11 +147,10 @@ fn u32assert2_fail() {
 
     expect_exec_error_matches!(
         test,
-        ExecutionError::NotU32Values{ values, err_code, label: _, source_file: _ } if
+        ExecutionError::NotU32Values{ values, label: _, source_file: _ } if
             values.len() == 2 &&
             values[0] == Felt::new(value_b) &&
-            values[1] == Felt::new(value_a) &&
-            err_code == ZERO
+            values[1] == Felt::new(value_a)
     );
 
     // -------- Case 2: a > 2^32 and b < 2^32 ---------------------------------------------------
@@ -162,10 +160,9 @@ fn u32assert2_fail() {
 
     expect_exec_error_matches!(
         test,
-        ExecutionError::NotU32Values{ values, err_code, label: _, source_file: _ } if
+        ExecutionError::NotU32Values{ values, label: _, source_file: _ } if
             values.len() == 1 &&
-            values[0] == Felt::new(value_a) &&
-            err_code == ZERO
+            values[0] == Felt::new(value_a)
     );
 
     // --------- Case 3: a < 2^32 and b > 2^32 --------------------------------------------------
@@ -175,10 +172,9 @@ fn u32assert2_fail() {
 
     expect_exec_error_matches!(
         test,
-        ExecutionError::NotU32Values{ values, err_code, label: _, source_file: _ } if
+        ExecutionError::NotU32Values{ values, label: _, source_file: _ } if
             values.len() == 1 &&
-            values[0] == Felt::new(value_b) &&
-            err_code == ZERO
+            values[0] == Felt::new(value_b)
     );
 }
 
@@ -206,10 +202,9 @@ fn u32assertw_fail() {
 
     expect_exec_error_matches!(
         test,
-        ExecutionError::NotU32Values{ values, err_code, label: _, source_file: _ } if
+        ExecutionError::NotU32Values{ values, label: _, source_file: _ } if
             values.len() == 2 &&
-            values.iter().all(|v| *v == Felt::new(U32_BOUND)) &&
-            err_code == ZERO
+            values.iter().all(|v| *v == Felt::new(U32_BOUND))
     );
 }
 
@@ -267,7 +262,7 @@ proptest! {
         let asm_op = "u32test";
 
         // check to see if the value of the element will be a valid u32
-        let expected_result = if value % Felt::MODULUS < U32_BOUND { 1 } else { 0 };
+        let expected_result = if value % Felt::ORDER_U64 < U32_BOUND { 1 } else { 0 };
 
         let test = build_op_test!(asm_op, &[value]);
         test.prop_expect_stack(&[expected_result, value])?;
@@ -318,7 +313,7 @@ proptest! {
 
         // expected result will be mod 2^32 applied to a field element
         // so the field modulus should be applied first
-        let expected_result = value % Felt::MODULUS % U32_BOUND;
+        let expected_result = value % Felt::ORDER_U64 % U32_BOUND;
 
         let test = build_op_test!(asm_op, &[value]);
         test.prop_expect_stack(&[expected_result])?;
@@ -330,7 +325,7 @@ proptest! {
 
         // expected result will be mod 2^32 applied to a field element
         // so the field modulus must be applied first
-        let felt_value = value % Felt::MODULUS;
+        let felt_value = value % Felt::ORDER_U64;
         let expected_b = felt_value >> 32;
         let expected_c = felt_value as u32 as u64;